refactor(phase-2e-strict): split OtpFill/OtpTrigger/OtpDetector to ≤10 LoC
Phase 2e-STRICT — drain the last 19 strict-overlay violations across the
three OTP Mediator clusters. Each monolithic file is split into co-located
sibling files behind a slim re-exporter so every function fits the strict
max-lines-per-function:10 ceiling and every file fits max-lines:150.

OtpFill (6 violations → 0):
  - OtpFillPhaseActions.PhoneHint.ts  — extractDeepPhoneHint (default)
  - OtpFillPhaseActions.Pre.ts        — executeFillPre (default, positional sig preserved)
  - OtpFillPhaseActions.Fill.ts       — executeFillAction (default, IRaceArgs bundle)
  - OtpFillPhaseActions.PostFinal.ts  — executeFillPost + executeFillFinal (named)
  - OtpFillPhaseActions.ts            — slim re-exporter

OtpTrigger (6 → 0):
  - OtpTriggerPhaseActions.Pre.ts     — executeTriggerPre (default, IFinalizePreArgs bundle)
  - OtpTriggerPhaseActions.Click.ts   — executeTriggerAction (default)
  - OtpTriggerPhaseActions.Post.ts    — executeTriggerPost (default, ICommitPostArgs bundle)
  - OtpTriggerPhaseActions.Final.ts   — executeTriggerFinal (default)
  - OtpTriggerPhaseActions.ts         — slim re-exporter

OtpDetector (7 → 0):
  - OtpDetector.Sequential.ts         — runSequential (default)
  - OtpDetector.Text.ts               — detectOtpScreen + extractPhoneHint (named)
  - OtpDetector.Submit.ts             — findOtpSubmitSelector (default)
  - OtpDetector.ClickText.ts          — tryClickTextInContexts (default)
  - OtpDetector.Click.ts              — clickOtpTriggerIfPresent + clickFromCandidates (named)
  - OtpDetector.ts                    — slim re-exporter preserving all 6 Common-shim symbols

Validation:
  - npx eslint src --max-warnings=0  → EXIT 0
  - npx tsc --noEmit                  → EXIT 0
  - npm run test:pipeline             → 5062/5062 pass

No production behaviour changes — pure restructure. Every public export
preserved (positional signatures, named-vs-default split, Common shim).
