+++ src/Scrapers/Pipeline/Mediator/Otp/OtpDetector.ts
@@ -1,445 +1,13 @@
-import { type Frame, type Page } from 'playwright-core';
-
-import type { SelectorCandidate } from '../../../Base/Config/LoginConfig.js';
-import type { LifecyclePromise } from '../../../Base/Interfaces/CallbackTypes.js';
-import { getDebug } from '../../Types/Debug.js';
-import { toXpathLiteral, tryInContext } from '../Selector/SelectorResolver.js';
-import {
-  OTP_FALLBACK_CLICK_TIMEOUT_MS,
-  OTP_FORCE_CLICK_TIMEOUT_MS,
-  OTP_INPUT_CANDIDATES,
-  OTP_SUBMIT_CANDIDATES,
-  OTP_TEXT_PATTERNS,
-  PHONE_PATTERN,
-  SMS_TRIGGER_CANDIDATES,
-} from './OtpDetectorConfig.js';
-
-const LOG = getDebug(import.meta.url);
-
-interface IRunState {
-  done: boolean;
-}
-
-/** Bundled args for {@link runStep} — under the 3-param ceiling. */
-interface IStepArgs<T> {
-  readonly state: IRunState;
-  readonly results: boolean[];
-  readonly action: (item: T, idx: number) => Promise<boolean>;
-}
-
 /**
- * Per-item step honouring the short-circuit flag in {@link IRunState}.
- * Extracted from {@link runSequential} so each function stays within
- * the per-function cap.
+ * OTP Detector orchestrator — slim re-exporter.
+ * Each detection/click concern lives in its own sibling file to
+ * satisfy the strict Mediator overlay (max-lines:150, max-lines-per-function:10).
  *
- * @param args - Bundle of short-circuit state, result slots, action.
- * @param item - Item to attempt.
- * @param idx - Position in the input array (used as result slot).
- * @returns Whether the action reported success for this item.
+ * Public surface preserved exactly so `src/Common/OtpDetector.ts`
+ * and every other consumer keeps importing the same six symbols.
  */
-async function runStep<T>(args: IStepArgs<T>, item: T, idx: number): Promise<boolean> {
-  if (args.state.done) return false;
-  const didSucceed = await args.action(item, idx);
-  args.results[idx] = didSucceed;
-  if (didSucceed) args.state.done = true;
-  return didSucceed;
-}
 
-/**
- * Build the reducer arrow consumed by {@link runSequential}'s
- * {@link Array.reduce}. Extracted so the chain factory call is not
- * inlined inside the reduce invocation (no-nested-call rule).
- *
- * @param args - Bundle of short-circuit state, result slots, action.
- * @returns Reducer accepted by `items.reduce<Promise<boolean>>(...)`.
- */
-function buildSequentialReducer<T>(
-  args: IStepArgs<T>,
-): (chain: Promise<boolean>, item: T, idx: number) => Promise<boolean> {
-  return (chain, item, idx) => chain.then(() => runStep(args, item, idx));
-}
-
-/**
- * Run actions sequentially, short-circuiting after the first success.
- * @param items - Array of items to process.
- * @param action - Async action returning true on success.
- * @returns Array of boolean results (true/false per item).
- */
-async function runSequential<T>(
-  items: T[],
-  action: (item: T, idx: number) => Promise<boolean>,
-): Promise<boolean[]> {
-  const state: IRunState = { done: false };
-  const results = items.map((): boolean => false);
-  const args: IStepArgs<T> = { state, results, action };
-  const reducer = buildSequentialReducer(args);
-  const seedChain = Promise.resolve(false);
-  await items.reduce<Promise<boolean>>((acc, item, idx) => reducer(acc, item, idx), seedChain);
-  return results;
-}
-
-export { OTP_SUBMIT_CANDIDATES };
-
-type TextCheckResult = 'otp' | 'clear' | 'unknown';
-
-/**
- * Retrieve the visible body text from the page for OTP pattern matching.
- * @param page - The Playwright page to read body text from.
- * @returns The body text content, or empty string on failure.
- */
-async function getBodyText(page: Page): Promise<string> {
-  try {
-    const text = await page.evaluate(() => document.body.innerText);
-    return typeof text === 'string' ? text : '';
-  } catch (error: unknown) {
-    LOG.debug(error, 'getBodyText failed (page context inaccessible)');
-    return '';
-  }
-}
-
-/**
- * Detect OTP presence by matching known text patterns in the page body.
- * @param page - The Playwright page to check.
- * @returns The detection result: 'otp', 'clear', or 'unknown'.
- */
-async function detectByText(page: Page): Promise<TextCheckResult> {
-  const bodyText = await getBodyText(page);
-  if (bodyText === '') return 'unknown';
-  return OTP_TEXT_PATTERNS.some(pattern => bodyText.includes(pattern)) ? 'otp' : 'clear';
-}
-
-/**
- * Check a single frame for OTP input candidates.
- * @param frame - The frame to search within.
- * @returns True if an OTP input was found in the frame.
- */
-async function checkFrameForOtpInput(frame: Frame): Promise<boolean> {
-  const found = await tryInContext(frame, OTP_INPUT_CANDIDATES);
-  return !!found;
-}
-
-/**
- * Detect OTP presence by searching for known OTP input fields on the page and in frames.
- * @param page - The Playwright page to search.
- * @param cachedFrames - Optional pre-filtered list of child frames.
- * @returns True if an OTP input field was found.
- */
-async function detectByInputField(page: Page, cachedFrames?: Frame[]): Promise<boolean> {
-  const found = await tryInContext(page, OTP_INPUT_CANDIDATES);
-  if (found) return true;
-  const mainFrame = page.mainFrame();
-  const frames = cachedFrames ?? page.frames().filter(f => f !== mainFrame);
-  const frameTasks = frames.map(checkFrameForOtpInput);
-  const results = await Promise.all(frameTasks);
-  return results.some(Boolean);
-}
-
-/**
- * Probe the page DOM for an OTP input field and log the outcome.
- * Wrapped here so {@link detectOtpScreen} stays a thin dispatcher
- * under the per-function cap.
- *
- * @param page - Playwright page to probe.
- * @returns True iff an OTP input field was found.
- */
-async function probeInputFieldAndLog(page: Page): Promise<boolean> {
-  const isByInput = await detectByInputField(page);
-  if (isByInput) LOG.debug('OTP detected by input field');
-  return isByInput;
-}
-
-/**
- * Log diagnostic for the 'unknown' text-result branch. Extracted so the
- * caller can stay a single-line conditional. See CR PR #286 F3 for why
- * 'unknown' does not short-circuit.
- *
- * @returns Sentinel `true` so the function has a meaningful return value
- *   (callers discard it — only the side-effect log matters).
- */
-function logTextUnknown(): true {
-  LOG.debug('Page text inaccessible — falling back to input-field probe');
-  return true;
-}
-
-/**
- * Detect whether the current page is showing an OTP screen.
- * @param page - The Playwright page to check for OTP indicators.
- * @returns True if an OTP screen is detected.
- */
-export async function detectOtpScreen(page: Page): Promise<boolean> {
-  const textResult = await detectByText(page);
-  if (textResult === 'otp') {
-    LOG.debug('OTP detected by text pattern');
-    return true;
-  }
-  if (textResult === 'unknown') logTextUnknown();
-  return probeInputFieldAndLog(page);
-}
-
-/**
- * Extract the masked phone number hint from the OTP screen body text.
- * @param page - The Playwright page to search for the phone hint.
- * @returns The matched phone hint string, or empty string if not found.
- */
-export async function extractPhoneHint(page: Page): Promise<string> {
-  const bodyText = await getBodyText(page);
-  const matched = PHONE_PATTERN.exec(bodyText);
-  return matched?.[0] ?? '';
-}
-
-/**
- * Find the Playwright selector for the OTP submit button across all frames.
- * @param page - The Playwright page to search for submit buttons.
- * @returns The matched selector (CSS or XPath), or empty string if not found.
- */
-export async function findOtpSubmitSelector(page: Page): Promise<string> {
-  const main = await tryInContext(page, OTP_SUBMIT_CANDIDATES);
-  if (main) return main;
-  const mainFrame = page.mainFrame();
-  const nonMainFrames = page.frames().filter(f => f !== mainFrame);
-  const frameTasks = nonMainFrames.map(frame => tryInContext(frame, OTP_SUBMIT_CANDIDATES));
-  const results = await Promise.all(frameTasks);
-  return results.find(sel => sel.length > 0) ?? '';
-}
-
-/**
- * Click the SMS trigger button if one is found on the page or in frames.
- * @param page - The Playwright page to search for SMS triggers.
- * @param cachedFrames - Optional pre-filtered list of child frames.
- * @returns True if a trigger was clicked, false if none found.
- */
-export async function clickOtpTriggerIfPresent(
-  page: Page,
-  cachedFrames?: Frame[],
-): Promise<boolean> {
-  const textValues = extractTextValues(SMS_TRIGGER_CANDIDATES);
-  const contexts = buildContextList(page, cachedFrames);
-  const didClick = await tryClickTextInContexts(contexts, textValues);
-  if (didClick) return true;
-  const hasFallback = await tryFallbackInAllContexts(contexts);
-  if (hasFallback) return true;
-  LOG.debug('No SMS trigger found — SMS may be auto-sent');
-  return false;
-}
-
-/**
- * Try SMS trigger fallback across all contexts sequentially.
- * @param contexts - Ordered list of Page/Frame contexts.
- * @returns True if a trigger was found and clicked.
- */
-async function tryFallbackInAllContexts(contexts: (Page | Frame)[]): Promise<boolean> {
-  const results = await runSequential(contexts, ctx =>
-    tryFallbackClick(ctx, SMS_TRIGGER_CANDIDATES),
-  );
-  return results.some(Boolean);
-}
-
-/** Text-based candidate kinds. */
-const TEXT_KINDS = ['textContent', 'clickableText'] as const;
-
-/**
- * Extract text values from text-based candidates.
- * @param candidates - Selector candidates to filter.
- * @returns Array of text values.
- */
-function extractTextValues(candidates: readonly SelectorCandidate[]): string[] {
-  return candidates
-    .filter(c => TEXT_KINDS.includes(c.kind as (typeof TEXT_KINDS)[number]))
-    .map(c => c.value);
-}
-
-/**
- * Try resolving and clicking a fallback selector in a single context.
- * @param ctx - The Page or Frame context.
- * @param candidates - Selector candidates to resolve.
- * @returns True if a selector was found and clicked.
- */
-async function tryFallbackClick(
-  ctx: Page | Frame,
-  candidates: readonly SelectorCandidate[],
-): Promise<boolean> {
-  const sel = await tryInContext(ctx, candidates);
-  if (!sel) return false;
-  LOG.debug('clickFromCandidates: fallback selector: %s', sel);
-  return ctx
-    .click(sel, { timeout: OTP_FALLBACK_CLICK_TIMEOUT_MS })
-    .then((): true => true)
-    .catch((): false => false);
-}
-
-/**
- * Try fallback selector resolution across all contexts sequentially.
- * @param contexts - Ordered list of Page/Frame contexts.
- * @param candidates - Selector candidates to resolve.
- * @returns True if a fallback selector was found and clicked.
- */
-async function tryFallbackInContexts(
-  contexts: (Page | Frame)[],
-  candidates: readonly SelectorCandidate[],
-): Promise<boolean> {
-  const results = await runSequential(contexts, ctx => tryFallbackClick(ctx, candidates));
-  return results.some(Boolean);
-}
-
-/**
- * Click the first matching candidate from a bank-specific selector list.
- * Uses the same tryInContext resolver pipeline as field resolution.
- * @param page - The Playwright page to search.
- * @param candidates - Ordered SelectorCandidate list (text-based preferred).
- * @param cachedFrames - Optional pre-filtered list of child frames.
- * @returns True if a candidate was found and clicked, false otherwise.
- */
-export async function clickFromCandidates(
-  page: Page,
-  candidates: readonly SelectorCandidate[],
-  cachedFrames?: Frame[],
-): Promise<boolean> {
-  const textValues = extractTextValues(candidates);
-  const contexts = buildContextList(page, cachedFrames);
-  const didClick = await tryClickTextInContexts(contexts, textValues);
-  if (didClick) return true;
-  const hasFallback = await tryFallbackInContexts(contexts, candidates);
-  if (hasFallback) return true;
-  LOG.debug('clickFromCandidates: no clickable match found');
-  return false;
-}
-
-/**
- * Build ordered list of contexts to search: main page then child frames.
- * @param page - The Playwright page.
- * @param cachedFrames - Optional cached frames list.
- * @returns Array of Page/Frame contexts to search.
- */
-function buildContextList(page: Page, cachedFrames?: Frame[]): (Page | Frame)[] {
-  const mainFrame = page.mainFrame();
-  const frames = cachedFrames ?? page.frames().filter(f => f !== mainFrame);
-  return [page, ...frames];
-}
-
-/**
- * Try clicking visible text in a single context.
- * @param ctx - Page or Frame to search.
- * @param texts - Text values to look for.
- * @param isMain - Whether ctx is the main page.
- * @returns True if a click succeeded.
- */
-async function tryClickTextInSingleContext(
-  ctx: Page | Frame,
-  texts: string[],
-  isMain: boolean,
-): Promise<boolean> {
-  const results = await sequentialClickAttempts(ctx, texts);
-  const idx = results.findIndex(Boolean);
-  if (idx >= 0) {
-    const label = isMain ? 'main' : 'frame';
-    LOG.debug('clicked "%s" in %s', texts[idx], label);
-    return true;
-  }
-  return false;
-}
-
-/**
- * Try clicking each text sequentially, stopping after first success.
- * @param ctx - Page or Frame to search.
- * @param texts - Text values to attempt.
- * @returns Array of results (true/false per text).
- */
-async function sequentialClickAttempts(ctx: Page | Frame, texts: string[]): Promise<boolean[]> {
-  return runSequential(texts, text => tryClickInnermostText(ctx, text));
-}
-
-/**
- * Try clicking visible text across all contexts.
- * @param contexts - Ordered list of contexts to search.
- * @param texts - Text values to look for.
- * @returns True if a click succeeded.
- */
-async function tryClickTextInContexts(
-  contexts: (Page | Frame)[],
-  texts: string[],
-): Promise<boolean> {
-  const results = await sequentialContextAttempts(contexts, texts);
-  const didMatch = results.some(Boolean);
-  if (!didMatch) {
-    LOG.debug('no match in %d contexts for %d texts', contexts.length, texts.length);
-  }
-  return didMatch;
-}
-
-/**
- * Try clicking text in each context sequentially, stopping after first success.
- * @param contexts - Ordered list of contexts.
- * @param texts - Text values to attempt.
- * @returns Array of results (true/false per context).
- */
-async function sequentialContextAttempts(
-  contexts: (Page | Frame)[],
-  texts: string[],
-): Promise<boolean[]> {
-  return runSequential(contexts, (ctx, idx) => tryClickTextInSingleContext(ctx, texts, idx === 0));
-}
-
-/** XPath predicate restricting matches to interactive elements. */
-const INTERACTIVE_FILTER = [
-  '(self::button or self::a or self::input',
-  'or self::select or self::textarea',
-  'or @role="button" or @role="link")',
-].join(' ');
-
-/**
- * Build an XPath matching the innermost interactive element with text.
- * @param text - The visible text to match.
- * @returns Playwright XPath selector for clickable elements only.
- */
-function innermostTextXpath(text: string): string {
-  const escaped = toXpathLiteral(text);
-  // CR PR #286 F8: only filter out the parent when an INTERACTIVE descendant
-  // contains the same text — otherwise patterns like
-  // `<button><span>אישור</span></button>` get incorrectly excluded because
-  // the span (a non-interactive descendant) contains the text.
-  return [
-    `xpath=//*[${INTERACTIVE_FILTER}`,
-    `and contains(., ${escaped})`,
-    `and not(.//*[${INTERACTIVE_FILTER} and contains(., ${escaped})])]`,
-  ].join(' ');
-}
-
-/**
- * Find the innermost element containing text and click it.
- * @param ctx - Page or Frame to search.
- * @param text - Visible text to find.
- * @returns True if clicked successfully.
- */
-async function tryClickInnermostText(ctx: Page | Frame, text: string): Promise<boolean> {
-  const xpathSelector = innermostTextXpath(text);
-  const locator = ctx.locator(xpathSelector);
-  const matches = await locator.all();
-  const results = await sequentialForceClicks(matches);
-  return results.some(Boolean);
-}
-
-/**
- * Try force-clicking matched elements sequentially, stopping after first success.
- * @param matches - Array of locator elements to attempt clicking.
- * @returns Array of click results (true/false per element).
- */
-async function sequentialForceClicks(matches: IClickable[]): Promise<boolean[]> {
-  return runSequential(matches, match => tryForceClick(match));
-}
-
-/** Locator with a click method. */
-interface IClickable {
-  click: (o: { timeout: number; force: boolean }) => LifecyclePromise;
-}
-
-/**
- * Attempt to force-click a single locator element.
- * @param loc - The Playwright locator to click.
- * @returns True if clicked, false on error.
- */
-async function tryForceClick(loc: IClickable): Promise<boolean> {
-  return loc
-    .click({ timeout: OTP_FORCE_CLICK_TIMEOUT_MS, force: true })
-    .then((): true => true)
-    .catch((): false => false);
-}
+export { clickFromCandidates, clickOtpTriggerIfPresent } from './OtpDetector.Click.js';
+export { default as findOtpSubmitSelector } from './OtpDetector.Submit.js';
+export { detectOtpScreen, extractPhoneHint } from './OtpDetector.Text.js';
+export { OTP_SUBMIT_CANDIDATES } from './OtpDetectorConfig.js';
